remove hdfs
