--- seatunnel-connectors/seatunnel-connector-spark-file/src/main/scala/org/apache/seatunnel/spark/sink/Hdfs.scala
@@ -1,33 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.seatunnel.spark.sink
-
-import org.apache.spark.sql.{Dataset, Row}
-import org.apache.seatunnel.common.config.CheckResult
-import org.apache.seatunnel.spark.Config.HDFS
-import org.apache.seatunnel.spark.SparkEnvironment
-
-class Hdfs extends FileSinkBase {
-
-  override def checkConfig(): CheckResult = {
-    checkConfigImpl(List(HDFS))
-  }
-
-  override def output(data: Dataset[Row], env: SparkEnvironment): Unit = {
-    outputImpl(data, HDFS)
-  }
-}
